--- Crexi.Models/Application.cs
@@ -1,10 +1,0 @@
-﻿
-namespace Crexi.Models;
-
-public class Application
-{
-    public Guid ApplicationId { get; set; }
-    public string? Name { get; set; }
-    public string? Description { get; set; }
-
-}
