--- Crexi.RulesEngine.Models/Enums/RulesEngineEnums.cs
@@ -1,12 +1,0 @@
-﻿namespace Crexi.RulesService.Models.Enums
-{
-    public enum ResponseCodeEnum
-    {
-        Success = 100,
-        SystemError = 500,
-        ValidationError = 600,
-        WorkflowNotFound = 700
-    }
-
-
-}
