--- Crexi.RulesEngine.Models/Requests/GetRulesResponse.cs
@@ -1,18 +1,0 @@
-﻿using Crexi.RulesService.Models.Enums;
-
-namespace Crexi.RulesService.Models.Requests;
-
-public class GetRulesResponse
-{
-    public ResponseCodeEnum ResponseCode { get; set; }
-    public string? ResponseMessage { get; set; }
-    public string? RuleUsed { get; set; }
-    public List<RulesResult> RulesResults{ get; set; }
-    public GetRulesResponse()
-    {
-        ResponseCode = ResponseCodeEnum.Success;
-        RulesResults = new();
-    }
-
-
-}
