--- Crexi.RulesEngine/Crexi.RulesService.csproj
@@ -1,18 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="RulesEngine" Version="5.0.3" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Crexi.RulesEngine.Models\Crexi.RulesService.Models.csproj" />
-    <ProjectReference Include="..\Crexi.Utilities\Crexi.Utilities.csproj" />
-  </ItemGroup>
-
-</Project>
