--- Crexi.RulesEngine/Interfaces/IRulesService.cs
@@ -1,8 +1,0 @@
-﻿using Crexi.RulesService.Models.Requests;
-
-namespace Crexi.RulesService.Interfaces;
-
-public interface IRulesService
-{
-    Task<GetRulesResponse> GetRulesAsync(GetRulesRequest request);
-}
