--- Crexi.RulesEngine/Services/RulesXsltService.cs
@@ -1,12 +1,0 @@
-﻿using Crexi.RulesService.Interfaces;
-using Crexi.RulesService.Models.Requests;
-
-namespace Crexi.RulesService.Services;
-
-public class RulesXsltService : IRulesService
-{
-    public Task<GetRulesResponse> GetRulesAsync(GetRulesRequest request)
-    {
-        throw new NotImplementedException();
-    }
-}
