--- Crexi.Utilities/Cache/Models/CacheOptions.cs
@@ -1,14 +1,0 @@
-﻿
-namespace Crexi.Cache.Models;
-
-public enum CacheExpiryOptionEnum : int
-{
-    Absolute,
-    Sliding
-}
-public class CacheOptions
-{
-    public CacheExpiryOptionEnum CacheExpiryOption { get; set; }
-    public double ExpiryTTLSeconds {get;set;}
-
-}
