+++ Models/Application.cs
@@ -1,0 +1,10 @@
+﻿
+namespace Models;
+
+public class Application
+{
+    public Guid ApplicationId { get; set; }
+    public string? Name { get; set; }
+    public string? Description { get; set; }
+
+}
