+++ Models/ClientApplication.cs
@@ -1,0 +1,14 @@
+﻿namespace Models;
+
+public class ClientApplication
+{
+    public Guid ClientApplicationId { get; set; }
+    public Guid ClientId { get; set; }
+    public Guid ApplicationId { get; set; }
+    public string? AuthToken { get; set; }
+    public DateTime TokenExpiration { get; set; }
+
+ 
+}
+
+
