+++ RateLimiter.sln
@@ -12,15 +12,15 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		README.md = README.md
 	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Crexi.Data.Models", "Crexi.Models\Crexi.Data.Models.csproj", "{C1A2CAED-42D3-4A70-ACC0-3717032EBF99}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Data.Models", "Models\Data.Models.csproj", "{C1A2CAED-42D3-4A70-ACC0-3717032EBF99}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RateLimiter.Models", "RateLimiter.Models\RateLimiter.Models.csproj", "{B6DAE4DC-9673-4D77-8674-D97C962F3E76}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Crexi.RulesService", "Crexi.RulesEngine\Crexi.RulesService.csproj", "{F49B7F6C-C531-432E-8076-0FA26FEEFA65}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RulesService", "RulesEngine\RulesService.csproj", "{F49B7F6C-C531-432E-8076-0FA26FEEFA65}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Crexi.RulesService.Models", "Crexi.RulesEngine.Models\Crexi.RulesService.Models.csproj", "{921C3DFC-57AA-4A27-97FD-89176B74A55C}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RulesService.Models", "RulesEngine.Models\RulesService.Models.csproj", "{921C3DFC-57AA-4A27-97FD-89176B74A55C}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Crexi.Utilities", "Crexi.Utilities\Crexi.Utilities.csproj", "{88BD8CDB-E94C-490F-B7F5-97B14BE46A29}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Utilities", "Utilities\Utilities.csproj", "{88BD8CDB-E94C-490F-B7F5-97B14BE46A29}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
