+++ RateLimiter/ServiceCollectionExtensions.cs
@@ -1,10 +1,10 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
-using Crexi.Cache.Providers;
+using Cache.Providers;
 using Microsoft.Extensions.Configuration;
 using RateLimiter.Services;
 using RateLimiter.Interfaces;
 
-namespace Crexi.Cache.Extensions;
+namespace Cache.Extensions;
 public static class ServiceCollectionExtensions
 {
     public static IServiceCollection ConfigureRateLimiter(this IServiceCollection services)
