+++ RulesEngine/Interfaces/IRulesService.cs
@@ -1,0 +1,8 @@
+﻿using RulesService.Models.Requests;
+
+namespace RulesService.Interfaces;
+
+public interface IRulesService
+{
+    Task<GetRulesResponse> GetRulesAsync(GetRulesRequest request);
+}
