+++ RulesEngine/Services/RulesXsltService.cs
@@ -1,0 +1,12 @@
+﻿using RulesService.Interfaces;
+using RulesService.Models.Requests;
+
+namespace RulesService.Services;
+
+public class RulesXsltService : IRulesService
+{
+    public Task<GetRulesResponse> GetRulesAsync(GetRulesRequest request)
+    {
+        throw new NotImplementedException();
+    }
+}
