+++ Utilities/Cache/Models/CacheOptions.cs
@@ -1,0 +1,14 @@
+﻿
+namespace Cache.Models;
+
+public enum CacheExpiryOptionEnum : int
+{
+    Absolute,
+    Sliding
+}
+public class CacheOptions
+{
+    public CacheExpiryOptionEnum CacheExpiryOption { get; set; }
+    public double ExpiryTTLSeconds {get;set;}
+
+}
